docs: Documentation Reorganization Foundations (#3815)
## What ❔

<!-- What are the changes this PR brings about? -->
We are establishing a foundation for reorganizing our documentation. 
In this pull request, several existing files have been repositioned to
align with the planned final structure.
The final structure of contracts-relevant documentation will be based on
[era-contracts](https://github.com/matter-labs/era-contracts/tree/main/docs)
repository.

## Why ❔
This reorganization will allow us to review the documentation relevant
to smart contracts and migrate it to the `zksync-era` repository, where
it will be accessible via a generated website.
Adjusting the file structure in the `zksync-era` repository is the first
step in this migration, and this PR begins that process.

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Co-authored-by: Marcin M <128217157+mm-zk@users.noreply.github.com>