docs: Add general contract and chain management docs (#3837)
## What ❔

<!-- What are the changes this PR brings about? -->
Continuation of the effort to reorganize our documentation.
In this pull request, general contract documentation
(README/overview/glossary) and chain management documentation have been
added.
Additionally, files containing `[TODO]` (indicating that they include
links that will need to be fixed once the files are finalized) have been
marked with `<!--- WIP --->` to denote that they are not yet complete
and are expected to change in the future.
This convention will also be followed in future PRs: when new files
containing `[TODO] `appear, they will be marked with the same comment.
The final structure of contracts-relevant documentation will be based on
[era-contracts](https://github.com/matter-labs/era-contracts/tree/main/docs)
repository.

## Why ❔
This reorganization will allow us to review the documentation relevant
to smart contracts and migrate it to the `zksync-era` repository, where
it will be accessible via a generated website.
Adjusting the file structure in the `zksync-era` repository is the first
step in this migration, and this PR begins that process.

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.