fix(contract-verifier): improve etherscan error handling (#4061)
## What ❔

**This PR improves Etherscan verificaation processing and error
handling.**

**Contract bytecode is not available error retry policy:**
It turned out that the retry policy for the `Contract bytecode is not
available` error doesn't work well, as it sometimes takes more than 10
minutes for Etherscan to index the contract data. This exceeds the total
retry timeout for the error.

To address this, I decided to introduce an increasing delay for this
specific error - starting at 2 minutes and going up to 1 day for the
final attempt.

IMO this approach is gonna work better: it begins with a quick retry to
handle the positive scenario, while still giving Etherscan enough time
to index the contract by using progressively longer timeouts for later
attempts.

Existing verifications that failed due to this error are requeued for
processing.

**Requests processing and error handling:**
Previously, in the case of any error, the verifier waited for a timeout
and retried the same request without attempting to process other
requests. This could be problematic, especially with the progressively
increasing timeouts introduced for `Contract bytecode is not available`
errors.

With this change, when a retryable error occurs, the verifier updates
the `retry_at` timestamp for the request and skips processing it until
the timeout has passed. In the meantime, it continues processing other
requests that are ready.

For indefinitely retryable errors, the verifier still pauses and does
not process other requests. These types of errors are typically not
specific to a particular request - they may indicate a service outage,
hitting a rate limit, a misconfiguration, etc. To avoid spamming
Etherscan, the verifier pauses and retries without moving on to other
queued requests.

**Metrics**
We lack visibility into when the Etherscan verification queue is
growing. I'm adding a metric for the number of queued requests so we can
set up proper alerts based on it.

## Why ❔

- To improve request processing and properly handle the `Contract
bytecode is not available` error.
- To improve service monitoring.

## Is this a breaking change?
- [ ] Yes
- [X] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [X] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.