feat(avail-client): async blob dispatch (#4010)
## What ❔

Enable async blob dispatching for the standard Avail client. 
In the previous implementation, `da_dispatcher` had to wait for the blob
to be finalized before sending another one.
In the current one, the blobs are only dispatched in `dispatch_blob`
function, but we don't wait for their finality there.
The finality check happens in the `ensure_finality`; if the deadline is
reached, the error will be returned, and the blob will be fully resent.

Remove `finality_state` config var.

## Why ❔

To improve the efficiency and fully utilize the `ensure_finality`.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.