feat(en): make `max_batches_to_recheck` configurable (#4082)
## What ❔

Make the number of batches to fetch by DA fetcher and to check by
ConsistencyChecker configurable.
The number can be the same for these two components because we only use
the DA information to create a correct commitment in the consistency
checker.


## Why ❔

The initial motivation was to avoid scenarios where ConsistencyChecker
can't progress for a very long because DA fetcher is fetching all the
info from batch 0. But I also noticed that we wanted to make it
configurable for ConsistencyChecker as well.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.