docs: Add interop typeA (cross-chain messaging) docs  (#4231)
## What ❔

<!-- What are the changes this PR brings about? -->
Continuation of the effort to reorganize our documentation.
In this pull request, interop typeA docs have been added.
Also, some of the links were fixed.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
This reorganization will allow us to review the documentation relevant
to smart contracts and migrate it to the zksync-era repository, where it
will be accessible via a generated website.

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.