feat(api): stabilize zks_gasPerPubdata (#4225)
## What ❔

This PR:
* Stabilizes `unstable_gasPerPubdata` as `zks_gasPerPubdata`
* Makes it never return less than `1` gas per pubdata as such
transactions are considered invalid
* Upgrades zksync-ethers to use a [WIP
branch](https://github.com/zksync-sdk/zksync-ethers/pull/252) that stops
calling `zks_estimateFee` in favour of `zks_gasPerPubdata` (if
available)

## Why ❔

`zks_estimateFee` is scheduled for deprecation and this is the last
missing piece

## Is this a breaking change?
- [x] Yes
- [ ] No

## Operational changes

This is technically a breaking change but realistically
`unstable_gasPerPubdata` is unused by the public

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.