feat: add `pubdata_limit` as batch parameter (#4228)
## What ❔

Adds `pubdata_limit` as batch parameter. It's sourced from config and
saved in DB for each batch as a part of `UnsealedBatchHeader`. It's
added to the consensus payload which is why it's only set for >= v29.

## Why ❔

Prerequisite for leader rotation: it's going to be used in seal
criterion instead of `config.max_pubdata_per_batch` so non-leader nodes
can verify that the batch fits into the limit. Currently, it's unused;
seal criteria will be modified separately.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
None

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.