chore: Remove legacy prover code (#4357)
## What ❔

Remove bincode encoding for prover-related types(ones involved in both
prover and core)
Remove API mode config option. 
So, after this PR it will be not possible to:
* Run proof data handler in server mode
* Run prover gateway in client mode
* Save/get files(like witness inputs/final proofs) from Object store
which are encoded with bincode

## Why ❔

This code was left in codebase for the sake of backwards compatibility.
Now, since these changes were introduced in protocol version 27, it is
"safe" to remove them in version 29.

## Is this a breaking change?
- [x] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.