fix(en): allow to disable sanity checks for commitment generation (#4487)
## What ❔

Disabling sanity checks for commitment generation


## Why ❔

Our external node is designed around commitment generation and
verification. Now it's possible to be in a situation, that we have
incorrect block and we can't revert it, because we don't have a
commitment for it.

EN is not able to produce any commitment => EN can't verify the
commitment => EN can't revert incorrect block.


## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

Signed-off-by: Danil <deniallugo@gmail.com>