feat(crypto): Update crypto deps (#4587)
## What ❔

Update crypto dependencies

Fix the issue with compilers. 
Ethereum foundation released a new compiler with 2 versions under the
same patch and it resulted an error inside svm-rs-build. The fix for the
library has been released under version 0.5.22, but it requires a new
rust. Simple upgrade to the newest nightly makes the crypto dependencies
not buildable.

I tried a few more version without a success. 

The best and easies way was hardcode the list of compilers and set the
env with.

But because we start the build from different places. It's required to
copy .cargo folder everywhere

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Signed-off-by: Danil <deniallugo@gmail.com>